moving functionality from `plot.js` to the data split file
